feat: enhance dark mode
