+++ src/share/components/is-dark-mode.ts
@@ -1,0 +1,22 @@
+import { prefs } from '@/share/core/storage';
+
+const isDarkMode = () => {
+  const darkMode = prefs.get('dark-mode');
+  switch (darkMode) {
+    case 'auto':
+      try {
+        const mql = window.matchMedia('(prefers-color-scheme: dark)');
+        return mql.matches;
+      } catch (e) {
+        // ignore
+      }
+      break;
+    case 'on':
+      return true;
+    default:
+      break;
+  }
+  return false;
+};
+
+export default isDarkMode;
